refactor: 게임 경기와 관련된 출력 부분 별도의 클래스 분리 작업 반영
