+++ src/test/java/step3/RacingGameTest.java
@@ -1,13 +1,8 @@
 package step3;
 
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import step3.domain.Car;
 import step3.domain.RacingGame;
-
-import java.util.ArrayList;
-import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -49,49 +44,5 @@ class RacingGameTest {
 
             assertThat(racingGame.getRound()).isEqualTo(whileCount);
         }
-    }
-
-    @ParameterizedTest
-    @CsvSource(value = {"kyd | 5", "kyd, one | 7", "kyd, one, two | 8"}, delimiter = '|')
-    void racingAll(String namesOfCarsText, int numberOfAttempts) {
-
-        RacingGame racingGame = new RacingGame(namesOfCarsText, numberOfAttempts);
-        racingGame.racingAll();
-
-        assertThat(racingGame.getRound()).isEqualTo(numberOfAttempts);
-    }
-
-    @Test
-    void champion() {
-        List<Car> cars = new ArrayList<>();
-
-        Car car1 = new Car("one");
-        car1.go(5);
-        car1.go(5);
-        cars.add(car1);
-
-        Car car2 = new Car("two");
-        car2.go(0);
-        car2.go(5);
-        cars.add(car2);
-
-        assertThat(RacingGame.getChampions(cars)).containsExactly(car1);
-    }
-
-    @Test
-    void championWithSamePosition() {
-        List<Car> cars = new ArrayList<>();
-
-        Car car1 = new Car("one");
-        car1.go(5);
-        car1.go(5);
-        cars.add(car1);
-
-        Car car2 = new Car("two");
-        car2.go(5);
-        car2.go(5);
-        cars.add(car2);
-
-        assertThat(RacingGame.getChampions(cars)).containsExactly(car1, car2);
     }
 }
